backend/requirements.txt
fastapi 
uvicorn
pydantic
sqlalchemy
pytz
python-dotenv
sqlalchemy 
pymysql
requests
weasyprint
fpdf
qrcode[pil]